Move EverestSplash to Celeste.Mod.mm and ditch separate project
