+++ Celeste.Mod.mm/Content/Graphics/Atlases/Gui/splash/bg_gradient_2x.png
