+++ Celeste.Mod.mm/Content/Graphics/Atlases/Gui/splash/everest_centered.png
