+++ Celeste.Mod.mm/Content/Graphics/Atlases/Gui/splash/splash_wheel_blur.png
