+++ Celeste.Mod.mm/Content/Graphics/Atlases/Gui/splash/starting_everest_text.png
