+++ Everest.sln
@@ -16,8 +16,6 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DiscordGameSDK", "DiscordGa
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NETCoreifier", "NETCoreifier\NETCoreifier.csproj", "{CD55DE23-DD59-4204-A556-C5DD1D032671}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EverestSplash", "EverestSplash\EverestSplash.csproj", "{47C6C795-188D-4F91-942E-ED44FED7F286}"
-EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -40,10 +38,6 @@ Global
 		{CD55DE23-DD59-4204-A556-C5DD1D032671}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{CD55DE23-DD59-4204-A556-C5DD1D032671}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{CD55DE23-DD59-4204-A556-C5DD1D032671}.Release|Any CPU.Build.0 = Release|Any CPU
-		{47C6C795-188D-4F91-942E-ED44FED7F286}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{47C6C795-188D-4F91-942E-ED44FED7F286}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{47C6C795-188D-4F91-942E-ED44FED7F286}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{47C6C795-188D-4F91-942E-ED44FED7F286}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
