--- EverestSplash/EverestSplash.csproj
@@ -1,17 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-  <PropertyGroup>
-    <TargetFramework>net7.0</TargetFramework>
-    <AssemblyName>EverestSplash</AssemblyName>
-    <RootNamespace>EverestSplash</RootNamespace>
-    <LangVersion>11</LangVersion>
-    <Nullable>enable</Nullable>
-    <OutputType>Library</OutputType>
-    <AllowUnsafeBlocks>true</AllowUnsafeBlocks>
-  </PropertyGroup>
-  
-  <ItemGroup>
-    <EmbeddedResource Include="SplashContent/**" />
-    <EmbeddedResource Include="../lib-stripped/Celeste-icon.png" />
-    <Folder Include="SDL2\" />
-  </ItemGroup>
-</Project>
